Fixed the file bunching on main, made routers
