+++ app/crud/user.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import Session
 from models.user import UserDB
-from schemas.user import UserCreate
+from schemas.user import UserCreate, User
 import bcrypt
 
 def create_user(db: Session, user: UserCreate):
@@ -16,6 +16,9 @@ def create_user(db: Session, user: UserCreate):
     db.commit()
     db.refresh(db_user)
     return db_user
+
+def update_user(db: Session, user: User):
+    pass
 
 def get_user_by_email(db: Session, email_address: str):
     return db.query(UserDB).filter(UserDB.email_address == email_address).first()
